Fixing nc.sh for Unix based systems
In Unix based OSes, which use bash, filesystems are generally case
sensitive. The nc.sh script on the other makes the assumption that
case insensitive filesystems, like Window's NTFS or FAT32 are used.
This also makes the nc.sh file executable by changing the file mode
from 0644 to 0755, this was likely due to the fact that this
program was developped using Windows which do not have a similar
file permission system.